requirement.txt
torch
torchvision
torchaudio
numpy
optuna
Pillow
matplotlib
fastapi
uvicorn